# docs/openapi/nestweb.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/exchange-code-for-user": {
      "post": {
        "operationId": "AuthController_exchangeCodeForUserId",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WechatCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/miniprogram-login": {
      "post": {
        "operationId": "AuthController_miniprogramLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WechatCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/validateCaptcha": {
      "post": {
        "operationId": "AuthController_validateCaptchaAndInitiateEmailVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpFormData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validate captcha and initiate email verification."
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/validateEmail": {
      "post": {
        "operationId": "AuthController_validateEmailToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validate email token and send SMS code."
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/validateSMS": {
      "post": {
        "operationId": "AuthController_validateSMSToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validate sms token and send Register"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/registerByEmail": {
      "post": {
        "operationId": "AuthController_registerByEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterByEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "验证邮箱验证码，若已存在用户则直接登录，否则创建新用户并登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "NestWeb Enterprise API",
    "description": "Enterprise admin backend API contract for Antdpro6.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "WechatCodeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "WeChat mini-program temporary login code."
          }
        },
        "required": [
          "code"
        ]
      },
      "Token": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "accessExpiresIn": {
            "type": "number"
          },
          "refreshExpiresIn": {
            "type": "number",
            "description": "Refresh token expiry timestamp in milliseconds. The refresh token itself is stored in an HttpOnly cookie."
          }
        },
        "required": [
          "accessToken",
          "accessExpiresIn",
          "refreshExpiresIn"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "SignUpFormData": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "captcha": {
            "type": "string"
          },
          "captchaToken": {
            "type": "string"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password",
          "confirmPassword",
          "country",
          "phoneNumber",
          "captcha",
          "captchaToken"
        ]
      },
      "ValidateTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "token",
          "code",
          "phone"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "email",
          "password",
          "phoneNumber"
        ]
      },
      "RegisterByEmailDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          }
        },
        "required": [
          "token",
          "code",
          "email",
          "password",
          "firstName",
          "lastName",
          "country",
          "phoneNumber"
        ]
      },
...
    }
  }
}